AULA04/Tarefa4/GatewaySolution/CommandModule/requirements.txt
smbus2
flask
google-cloud-pubsub